feat: Extract custom icon deletion logic to utility function (i906)
This commit refactors the custom icon deletion logic by extracting it
into a new internal utility function, `deleteGridItemCustomIconFile`.
This function encapsulates the file deletion process for various
`GridItemData` types.

The `DeleteGridItemCustomIconUseCase` and
`UpdateGridItemCustomIconUseCase` now leverage this utility function.
Additionally, `UpdateEblanApplicationInfoCustomIconUseCase` also uses
this function to clean up old custom icon files before saving new ones.

This change promotes code reuse and simplifies the implementation of
use cases related to custom icon management.

Affected files:
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/GridUtil.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/application/UpdateEblanApplicationInfoCustomIconUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/DeleteGridItemCustomIconUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/UpdateGridItemCustomIconUseCase.kt
